Refactor process alive check in CloudServerServiceFactory
This commit simplifies the check for whether or not a cloud service process is alive in CloudServerServiceFactory. It also removes an unused import statement related to the CloudServiceShutdownPacket. Now if the process is null, it directly returns false. If the process exists, the rest of the checking logic is the same as before.